+++ src/app/pages/calendar/calendar.component.ts
@@ -14,7 +14,7 @@ import { CalendarEventsService } from '@services/calendar-events.service';
   selector: 'app-calendar',
   changeDetection: ChangeDetectionStrategy.OnPush,
   templateUrl: './calendar.component.html',
-  styleUrls: ['./calendar.component.css']
+  styleUrls: ['./calendar.component.scss']
 })
 export class CalendarComponent implements OnInit {